requirements.txt
attrs==20.3.0
aws-cfn-bootstrap==2.0
awscli==2.23.11
awscrt==0.23.8
Babel==2.9.1
blinker==1.9.0
cffi==1.14.5
chardet==4.0.0
chevron==0.13.1
click==8.1.8
cloud-init==22.2.2
colorama==0.4.4
configobj==5.0.6
cryptography==36.0.1
dbus-python==1.2.18
distlib==0.3.9
distro==1.5.0
docutils==0.16
ec2-hibinit-agent==1.0.8
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
gpg==1.15.1
greenlet==3.1.1
gunicorn==23.0.0
idna==2.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==0.10.0
jsonpatch==1.21
jsonpointer==2.0
jsonschema==3.2.0
libcomps==0.1.20
lockfile==0.12.2
MarkupSafe==3.0.2
netifaces==0.10.6
oauthlib==3.0.2
packaging==24.2
platformdirs==4.3.7
ply==3.11
prettytable==0.7.2
prompt-toolkit==3.0.24
pycparser==2.20
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.8.1
python-dotenv==1.1.0
pytz==2022.7.1
PyYAML==5.4.1
release-notification==1.2
requests==2.25.1
rpm==4.16.1.3
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
selinux==3.4
sepolicy==3.4
setools==4.4.1
six==1.15.0
SQLAlchemy==2.0.40
support-info==1.0
systemd-python==235
typing_extensions==4.13.1
urllib3==1.25.10
virtualenv==20.29.3
wcwidth==0.2.5
Werkzeug==3.1.3
zipp==3.21.0
